feat: fix cast long to biginteger for stub request;

+++ mockingbird-aggregator/src/main/java/org/qubership/automation/itf/integration/TriggerExecutor.java
@@ -437,7 +437,7 @@ public class TriggerExecutor implements IDiameterEventProducer {
             log.warn("More than one environment with inbound system {} and server {} found", system, server);
         }
         Map<String, Object> map = new HashMap<>();
-        map.put("id", environmentInfo.getFirst()[0]);
+        map.put("id", BigInteger.valueOf((Long) environmentInfo.getFirst()[0]));
         map.put("name", environmentInfo.getFirst()[1]);
         return map;
     }
